refactor: move user module to lib/zero/user/
Part of the lib/ restructuring epic (#7697). Moves user lifecycle
services (deletion, external cleanup, S3 cleanup, preferences) into
the zero/ layer where they belong.

Closes #7732
