+++ turbo/apps/web/src/lib/zero/user/user-deletion-service.ts
@@ -1,28 +1,28 @@
 import { eq, and, inArray } from "drizzle-orm";
-import { logger } from "../logger";
-import { agentRuns } from "../../db/schema/agent-run";
-import { agentRunQueue } from "../../db/schema/agent-run-queue";
-import { agentComposes } from "../../db/schema/agent-compose";
-import { storages } from "../../db/schema/storage";
-import { secrets } from "../../db/schema/secret";
-import { modelProviders } from "../../db/schema/model-provider";
-import { connectors } from "../../db/schema/connector";
-import { variables } from "../../db/schema/variable";
-import { usageDaily } from "../../db/schema/usage-daily";
-import { exportJobs } from "../../db/schema/export-job";
-import { zeroAgentSchedules } from "../../db/schema/zero-agent-schedule";
-import { slackOrgConnections } from "../../db/schema/slack-org-connection";
-import { slackOrgPendingQuestions } from "../../db/schema/slack-org-pending-question";
-import { githubUserLinks } from "../../db/schema/github-user-link";
-import { telegramUserLinks } from "../../db/schema/telegram-user-link";
-import { cliTokens } from "../../db/schema/cli-tokens";
-import { composeJobs } from "../../db/schema/compose-job";
-import { connectorSessions } from "../../db/schema/connector-session";
-import { deviceCodes } from "../../db/schema/device-codes";
-import { orgMembersCache } from "../../db/schema/org-members-cache";
-import { orgMembersMetadata } from "../../db/schema/org-members-metadata";
-import { userCache } from "../../db/schema/user-cache";
-import { users } from "../../db/schema/user";
+import { logger } from "../../logger";
+import { agentRuns } from "../../../db/schema/agent-run";
+import { agentRunQueue } from "../../../db/schema/agent-run-queue";
+import { agentComposes } from "../../../db/schema/agent-compose";
+import { storages } from "../../../db/schema/storage";
+import { secrets } from "../../../db/schema/secret";
+import { modelProviders } from "../../../db/schema/model-provider";
+import { connectors } from "../../../db/schema/connector";
+import { variables } from "../../../db/schema/variable";
+import { usageDaily } from "../../../db/schema/usage-daily";
+import { exportJobs } from "../../../db/schema/export-job";
+import { zeroAgentSchedules } from "../../../db/schema/zero-agent-schedule";
+import { slackOrgConnections } from "../../../db/schema/slack-org-connection";
+import { slackOrgPendingQuestions } from "../../../db/schema/slack-org-pending-question";
+import { githubUserLinks } from "../../../db/schema/github-user-link";
+import { telegramUserLinks } from "../../../db/schema/telegram-user-link";
+import { cliTokens } from "../../../db/schema/cli-tokens";
+import { composeJobs } from "../../../db/schema/compose-job";
+import { connectorSessions } from "../../../db/schema/connector-session";
+import { deviceCodes } from "../../../db/schema/device-codes";
+import { orgMembersCache } from "../../../db/schema/org-members-cache";
+import { orgMembersMetadata } from "../../../db/schema/org-members-metadata";
+import { userCache } from "../../../db/schema/user-cache";
+import { users } from "../../../db/schema/user";
 
 const log = logger("service:user-deletion");
