+++ turbo/apps/web/src/lib/zero/user/user-external-cleanup.ts
@@ -1,12 +1,12 @@
 import { eq } from "drizzle-orm";
 import { connectorTypeSchema } from "@vm0/core";
-import { logger } from "../logger";
-import { revokeConnectorToken } from "../connector/connector-service";
-import { connectors } from "../../db/schema/connector";
-import { githubUserLinks } from "../../db/schema/github-user-link";
-import { telegramUserLinks } from "../../db/schema/telegram-user-link";
-import { slackOrgConnections } from "../../db/schema/slack-org-connection";
-import { slackOrgPendingQuestions } from "../../db/schema/slack-org-pending-question";
+import { logger } from "../../logger";
+import { revokeConnectorToken } from "../../connector/connector-service";
+import { connectors } from "../../../db/schema/connector";
+import { githubUserLinks } from "../../../db/schema/github-user-link";
+import { telegramUserLinks } from "../../../db/schema/telegram-user-link";
+import { slackOrgConnections } from "../../../db/schema/slack-org-connection";
+import { slackOrgPendingQuestions } from "../../../db/schema/slack-org-pending-question";
 
 const log = logger("service:user-external-cleanup");
