+++ turbo/apps/web/src/lib/zero/user/user-s3-cleanup.ts
@@ -1,8 +1,8 @@
 import { eq, and, isNotNull } from "drizzle-orm";
-import { storages } from "../../db/schema/storage";
-import { exportJobs } from "../../db/schema/export-job";
-import { listS3Objects, deleteS3Objects } from "../s3/s3-client";
-import { logger } from "../logger";
+import { storages } from "../../../db/schema/storage";
+import { exportJobs } from "../../../db/schema/export-job";
+import { listS3Objects, deleteS3Objects } from "../../s3/s3-client";
+import { logger } from "../../logger";
 
 const log = logger("service:user-s3-cleanup");
